reports/requirements.txt
Flask==3.0.0
python-dotenv==0.21.1
Flask-JWT-Extended==4.5.2
Cerberus==1.3.4
bcrypt==4.0.1
Flask-Cors==3.0.10
PyMySQL==1.1.0